# Conflicts:
#	Tower Defense/Scenes/Components/Projectile/ProjectileInfo.cs
#	Tower Defense/Scripts/GameData.cs

+++ Tower Defense/Scenes/Components/Projectile/ProjectileInfo.cs
@@ -4,6 +4,6 @@ using System;
 [GlobalClass]
 public partial class ProjectileInfo : Resource
 {
-	[Export] public float Speed;
+	[Export] public int Speed;
 	[Export] public int Damage;
 }
